# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent-eBPF MCP Gateway",
    "description": "Production-Grade Security Gateway with eBPF Integration",
    "version": "2.0.0-ULTRA"
  },
  "paths": {
...
    "/api/v1/policy/compile": {
      "post": {
        "tags": [
          "NLP Policy Compiler"
        ],
        "summary": "Compile Policy From Text",
        "description": "Translates natural language security rules into typed eBPF/Cgroup policies via Gemini AI.",
        "operationId": "compile_policy_from_text_api_v1_policy_compile_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompiledEbpfRule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compiler/compile": {
      "post": {
        "tags": [
          "NLP Policy Compiler"
        ],
        "summary": "Compile Policy From Nlp",
        "description": "Synchronous fallback endpoint for compiler evaluations.",
        "operationId": "compile_policy_from_nlp_api_v1_compiler_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyCompilationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compiler/apply": {
      "post": {
        "tags": [
          "NLP Policy Compiler"
        ],
        "summary": "Apply Compiled Rule",
        "description": "Directly provisions a compiled natural language rule into the active eBPF kernel policy map.",
        "operationId": "apply_compiled_rule_api_v1_compiler_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyCompiledRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/billing/usage": {
      "get": {
        "tags": [
          "SaaS Billing"
        ],
        "summary": "Api Get Billing Usage",
        "description": "Returns instantaneous billing usage, event counters, and quota progress.",
        "operationId": "api_get_billing_usage_api_v1_billing_usage_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default-tenant",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/checkout": {
      "post": {
        "tags": [
          "SaaS Billing"
        ],
        "summary": "Api Create Checkout",
        "description": "Creates Stripe checkout session for Team Pro ($99/mo) or Enterprise ($499/mo).",
        "operationId": "api_create_checkout_api_v1_billing_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register Endpoint",
        "operationId": "register_endpoint_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify Otp Endpoint",
        "operationId": "verify_otp_endpoint_api_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/resend-otp": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Resend Otp Endpoint",
        "operationId": "resend_otp_endpoint_api_auth_resend_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login Endpoint",
        "operationId": "login_endpoint_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyCompiledRuleRequest": {
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/CompiledRule"
          },
          "pid": {
            "type": "integer",
            "title": "Pid",
            "description": "Optional target process ID to immediately bind the rule",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "rule"
        ],
        "title": "ApplyCompiledRuleRequest"
      },
...
      "CheckoutRequest": {
        "properties": {
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "plan_tier": {
            "type": "string",
            "title": "Plan Tier",
            "default": "team_pro"
          },
          "success_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Url"
          },
          "cancel_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "title": "CheckoutRequest"
      },
...
      "CompileRequest": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Natural language policy prompt"
          },
          "natural_language_rule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Natural Language Rule",
            "description": "Natural language security intent rule"
          }
        },
        "type": "object",
        "title": "CompileRequest"
      },
      "CompiledEbpfRule": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id",
            "description": "Unique policy identifier (e.g. pol-net-01)"
          },
          "rule_name": {
            "type": "string",
            "title": "Rule Name",
            "description": "Short summary description of the rule"
          },
          "target_comm": {
            "type": "string",
            "title": "Target Comm",
            "description": "Target process executable or binary name (e.g. python3, nc, curl)"
          },
          "allowed_ports": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Allowed Ports",
            "description": "List of permitted network ports"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "TCP",
              "UDP",
              "ALL"
            ],
            "title": "Protocol",
            "description": "Target transport layer protocol",
            "default": "TCP"
          },
          "action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "DROP",
              "LOG_ONLY"
            ],
            "title": "Action",
            "description": "Enforcement action verdict",
            "default": "ALLOW"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Risk Level",
            "description": "AI evaluated threat risk level",
            "default": "LOW"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "description": "Security rationale provided by the AI compiler"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "rule_name",
          "target_comm",
          "rationale"
        ],
        "title": "CompiledEbpfRule"
      },
      "CompiledRule": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_comm": {
            "type": "string",
            "title": "Target Comm"
          },
          "allowed_ports": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Allowed Ports"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "target_comm",
          "allowed_ports",
          "action",
          "risk_score",
          "explanation"
        ],
        "title": "CompiledRule"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "PolicyCompilationResult": {
        "properties": {
          "natural_query": {
            "type": "string",
            "title": "Natural Query"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/CompiledRule"
            },
            "type": "array",
            "title": "Rules"
          },
          "overall_safety_rating": {
            "type": "string",
            "title": "Overall Safety Rating"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "natural_query",
          "rules",
          "overall_safety_rating"
        ],
        "title": "PolicyCompilationResult"
      },
...
      "RegisterRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name",
            "default": "Autonomous AI Corp"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "ResendOtpRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ResendOtpRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyOtpRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "otp_code": {
            "type": "string",
            "title": "Otp Code"
          }
        },
        "type": "object",
        "required": [
          "email",
          "otp_code"
        ],
        "title": "VerifyOtpRequest"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}